Google Analytics work around
Google analytics is not tracking user activity.
Based in a brief research, this is related to the version of GA API
being called from Ga plugin.

New GA codies are using V2 but the pluging calls V1.

In an attempt to fix this, I'm adding a loca ga-temp.js and using it
instead of CDN oficial version

Not sure if this will solve the issue but worth the attempt

You can find more information about the issue here:
https://github.com/docsifyjs/docsify/issues/1695